requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.5.0
matplotlib>=3.7.1
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pillow>=10.0.0 